# Conflicts:
#	TFG_server/src/main/java/org/server/dao/repositories/UserRepository.java
#	TFG_server/src/main/java/org/server/ui/controllers/NotasController.java

+++ TFG_server/src/main/java/org/server/domain/service/UserService.java
@@ -27,7 +27,7 @@ public class UserService {
                 .toList();
     }
 
-    public boolean addNoteToSaved(String username, Long noteId) {
+    public boolean addNoteToSaved(String username, int noteId) {
         boolean alreadyExists = userSavedRepository.findByUserUsername(username).stream()
                 .anyMatch(savedNote -> savedNote.getNote().getId() == (noteId));
